pyproject.toml
[tool.poetry]
name = "clogged"
version = "0.1.0"
description = ""
authors = ["Max Smagin <lazlogogalac777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
argon2-cffi = "^23.1.0"
asyncpg = "^0.30.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
redis = "^5.1.1"

[tool.poetry.scripts]
start = "clogged.start:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
